move schema into /types, edit DAO to use Query enum
